pyproject.toml
[project]
name = "pdf-extractext"
version = "0.1.0"
description = "Convertidor de PDF"
readme = "README.md"
requires-python = "==3.14.4"

dependencies = [
    "fastapi==0.135.3",
    "pydantic==2.13.1",
    "pydantic-settings==2.13.1",
    "python-multipart==0.0.26",
    "uvicorn[standard]==0.44.0",
    "python-dotenv==1.2.2",
    "pypdf==6.10.1",
    "pymongo==4.17.0"
    
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.3",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.1.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "reportlab==4.4.10",
    "func-timeout==4.3.5",
    "mongomock==4.3.0"
    
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
